requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
httpx
cachetools
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
